split out a graph file, add benchmark for large dependency graph
